decouple agents to keep them working over time by duplicating a bit of code
