+++ nidup/pysc2/agent/hybrid/army_learning.py
@@ -1,10 +1,8 @@
 
-import math
 import numpy as np
 from nidup.pysc2.agent.order import Order
 from nidup.pysc2.agent.information import Location, BuildingCounter, MinimapEnemyHotSquaresBuilder, EnemyDetector, RaceNames
-from nidup.pysc2.agent.smart.orders import BuildMarine, BuildMarauder, NoOrder
-from nidup.pysc2.agent.hybrid.attack_order_learning import QLearningAttack
+from nidup.pysc2.agent.hybrid.orders import BuildMarine, BuildMarauder, NoOrder, DumbAttack
 from nidup.pysc2.wrapper.observations import Observations
 
 ACTION_DO_NOTHING = 'donothing'
@@ -44,7 +42,7 @@ class SmartActions:
         elif smart_action == ACTION_BUILD_MARAUDER:
             return BuildMarauder(self.location)
         elif smart_action == ACTION_ATTACK:
-            return QLearningAttack(self.location, int(x), int(y))
+            return DumbAttack(self.location, int(x), int(y))
         elif smart_action == ACTION_DO_NOTHING:
             return NoOrder()
         else:
