+++ nidup/pysc2/agent/smart/commander.py
@@ -6,7 +6,7 @@ from nidup.pysc2.agent.order import Order
 from nidup.pysc2.learning.qlearning import QLearningTable, QLearningTableStorage
 from nidup.pysc2.wrapper.observations import Observations
 from nidup.pysc2.agent.information import Location
-from nidup.pysc2.agent.smart.orders import BuildBarrack, BuildSupplyDepot, BuildMarine, DumbAttack, NoOrder, PrepareSCVControlGroupsOrder
+from nidup.pysc2.agent.smart.orders import BuildBarracks, BuildSupplyDepot, BuildMarine, Attack, NoOrder
 from nidup.pysc2.wrapper.unit_types import UnitTypeIds
 
 _PLAYER_SELF = 1
@@ -48,14 +48,14 @@ class SmartActions:
         smart_action, x, y = self._split_action(action_id)
         if smart_action == ACTION_BUILD_BARRACKS:
             max_barracks = 2
-            return BuildBarrack(self.location, max_barracks)
+            return BuildBarracks(self.location, max_barracks)
         elif smart_action == ACTION_BUILD_SUPPLY_DEPOT:
             max_supplies = 2
             return BuildSupplyDepot(self.location, max_supplies)
         elif smart_action == ACTION_BUILD_MARINE:
             return BuildMarine(self.location)
         elif smart_action == ACTION_ATTACK:
-            return DumbAttack(self.location, int(x), int(y))
+            return Attack(self.location, int(x), int(y))
         elif smart_action == ACTION_DO_NOTHING:
             return NoOrder()
         else:
@@ -116,7 +116,6 @@ class QLearningCommander(Commander):
         self.previous_state = None
         self.previous_order = None
         self.location = None
-        self.control_group_order = None
 
     def order(self, observations: Observations, step_index: int)-> Order:
         if observations.last():
@@ -134,12 +133,8 @@ class QLearningCommander(Commander):
             self.smart_actions = SmartActions(self.location)
             self.qlearn = QLearningTable(actions=list(range(len(self.smart_actions.all()))))
             QLearningTableStorage().load(self.qlearn, self.agent_name)
-            self.control_group_order = PrepareSCVControlGroupsOrder(self.location)
 
-        if not self.control_group_order.done(observations):
-            return self.control_group_order
-
-        elif not self.previous_order or self.previous_order.done(observations):
+        if not self.previous_order or self.previous_order.done(observations):
 
             current_state = StateBuilder().build_state(self.location, observations)
