+++ nidup/pysc2/agent/smart/orders.py
@@ -2,7 +2,7 @@
 import random
 from pysc2.lib import actions
 from nidup.pysc2.agent.order import Order
-from nidup.pysc2.agent.information import Location, BuildingCounter
+from nidup.pysc2.agent.information import Location
 from nidup.pysc2.wrapper.actions import TerranActions, TerranActionIds
 from nidup.pysc2.wrapper.observations import Observations
 from nidup.pysc2.wrapper.unit_types import UnitTypeIds
@@ -16,157 +16,11 @@ class SmartOrder(Order):
         self.action_ids = TerranActionIds()
         self.unit_type_ids = UnitTypeIds()
 
-    def doable(self, observations: Observations) -> bool:
-        raise NotImplementedError("Should be implemented by concrete order")
-
     def done(self, observations: Observations) -> bool:
         raise NotImplementedError("Should be implemented by concrete order")
 
     def execute(self, observations: Observations) -> actions.FunctionCall:
         raise NotImplementedError("Should be implemented by concrete order")
-
-
-class SCVControlGroups:
-
-    def all_group_id(self) -> int:
-        return 0
-
-    def mineral_collectors_group_id(self) -> int:
-        return 1
-
-    def refinery_one_collectors_group_id(self) -> int:
-        return 2
-
-    def refinery_two_collectors_group_id(self) -> int:
-        return 3
-
-
-class BuildingPositionsFromCommandCenter:
-
-    def supply_depots(self) -> []:
-        # some tweak to make it work on both start positions that seems not symetric
-        return [
-            [-35, -20],
-            [-35, -10],
-            [-35, 0],
-            [-35, 10],
-            [-35, 20],
-            [-25, -30],
-            [-15, -30],
-            [-5, -30],
-            [5, -35],
-            [15, -35]
-        ]
-
-    # keeping space for techlab/reactor on each
-    def barracks(self) -> []:
-        return [
-            [15, -10],
-            [30, -20],
-            [20, 10],
-            [30, 20],
-        ]
-
-    def factories(self) -> []:
-        return [
-            [10, 25],
-            [30, 25], # check it, could not work
-        ]
-
-    def vespene_geysers(self) -> []:
-        return [
-            [10, -22],
-            [-24, 12]
-        ]
-
-    def refineries(self) -> []:
-        return self.vespene_geysers()
-
-
-# Groups all scv to specialized group to facilitate the further selections of dedicated workers, at the end of this
-# order, control groups look like the following
-# print(observations.control_groups())
-# [[45 12] <- all scv
-#  [45  6] <- mineral scv
-#  [45  3] <- vespene1 scv
-#  [45  3] <- vespene2 scv
-#  [0  0]  <- free control group
-#  [0  0]
-#  [0  0]
-#  [0  0]
-#  [0  0]
-#  [0  0]]
-class PrepareSCVControlGroupsOrder(SmartOrder):
-
-    def __init__(self, location: Location):
-        SmartOrder.__init__(self, location)
-        self.step = 1
-        groups = SCVControlGroups()
-        self.all_scv_group_id = groups.all_group_id()
-        self.mineral_group_id = groups.mineral_collectors_group_id()
-        self.vespene_group1_id = groups.refinery_one_collectors_group_id()
-        self.vespene_group2_id = groups.refinery_two_collectors_group_id()
-        self.expected_group_sizes = {
-            self.mineral_group_id: 6,
-            self.vespene_group1_id: 3,
-            self.vespene_group2_id: 3,
-        }
-        self.scv_index_in_all_group = 0
-        self.current_group_index = 1
-
-    def doable(self, observations: Observations) -> bool:
-        return True
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 6
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        if self.step == 1:
-            return self._select_all_vcs()
-        elif self.step == 2:
-            return self._add_selected_vcs_to_the_all_group()
-        # we loop on the following 3 actions to prepare all specialized groups
-        elif self.step == 3:
-            return self._select_the_all_group()
-        elif self.step == 4:
-            return self._select_a_new_vcs_from_all_group(observations)
-        elif self.step == 5:
-            return self._add_selected_vcs_to_a_specialized_group(self.current_group_index)
-        return self.actions.no_op()
-
-    def _select_all_vcs(self) -> actions.FunctionCall:
-        self.step = self.step + 1
-        return self.actions.select_rect([0, 0], [83, 83])
-
-    def _add_selected_vcs_to_the_all_group(self) -> actions.FunctionCall:
-        self.step = self.step + 1
-        return self.actions.add_control_group(self.all_scv_group_id)
-
-    def _select_the_all_group(self) -> actions.FunctionCall:
-        self.step = self.step + 1
-        return self.actions.select_control_group(self.all_scv_group_id)
-
-    def _select_a_new_vcs_from_all_group(self, observations: Observations) -> actions.FunctionCall:
-        if self.action_ids.select_unit() in observations.available_actions():
-            self.step = self.step + 1
-            action = self.actions.select_unit(self.scv_index_in_all_group)
-            self.scv_index_in_all_group = self.scv_index_in_all_group + 1
-            return action
-
-    def _add_selected_vcs_to_a_specialized_group(self, group_id: int) -> actions.FunctionCall:
-        self.step = self.step + 1
-        self._goto_next_vcs_to_group(group_id)
-        return self.actions.add_control_group(group_id)
-
-    def _goto_next_vcs_to_group(self, group_id: int):
-        # fulfill the current group
-        if self.expected_group_sizes[group_id] > 1:
-            self.expected_group_sizes[group_id] = self.expected_group_sizes[group_id] - 1
-            self.step = self.step - 3
-        # move  to the next group
-        elif self.current_group_index < len(self.expected_group_sizes):
-            self.current_group_index = self.current_group_index + 1
-            self.step = self.step - 3
 
 
 class SCVCommonActions:
@@ -176,8 +30,15 @@ class SCVCommonActions:
         self.action_ids = TerranActionIds()
         self.unit_type_ids = UnitTypeIds()
 
-    def select_a_group_of_scv(self, group_id: int) -> actions.FunctionCall:
-        return self.actions.select_control_group(group_id)
+    def select_scv(self, observations: Observations) -> actions.FunctionCall:
+        unit_type = observations.screen().unit_type()
+        unit_y, unit_x = (unit_type == self.unit_type_ids.terran_scv()).nonzero()
+        if unit_y.any():
+            i = random.randint(0, len(unit_y) - 1)
+            target = [unit_x[i], unit_y[i]]
+            return self.actions.select_point(target)
+
+        return self.actions.no_op()
 
     def send_scv_to_mineral(self, observations: Observations) -> actions.FunctionCall:
         if self.action_ids.harvest_gather() in observations.available_actions():
@@ -191,54 +52,42 @@ class SCVCommonActions:
                 return self.actions.harvest_gather(target)
         return self.actions.no_op()
 
-    def send_selected_scv_group_to_refinery(self, location: Location, observations: Observations, refinery_id: int) -> actions.FunctionCall:
-        if self.action_ids.harvest_gather() in observations.available_actions():
-            if refinery_id == 1:
-                difference_from_cc = BuildingPositionsFromCommandCenter().vespene_geysers()[0]
-            else:
-                difference_from_cc = BuildingPositionsFromCommandCenter().vespene_geysers()[1]
-            cc_y, cc_x = location.command_center_first_position()
-            target = location.transform_distance(
-                round(cc_x.mean()),
-                difference_from_cc[0],
-                round(cc_y.mean()),
-                difference_from_cc[1],
-            )
-            #print("send collector " + str(refinery_id) + " to refinery " + str(target[0]) + " " + str(target[1]))
-            return self.actions.harvest_gather(target)
-        return self.actions.no_op()
 
-
-class BuildBarrack(SmartOrder):
+class BuildBarracks(SmartOrder):
 
     def __init__(self, location: Location, max_barracks: int = 4):
         SmartOrder.__init__(self, location)
         self.step = 0
-        self.scv_groups = SCVControlGroups()
         self.max_barracks = max_barracks
 
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 150
-
     def done(self, observations: Observations) -> bool:
-        return self.step == 2
+        return self.step == 3
 
     def execute(self, observations: Observations) -> actions.FunctionCall:
         self.step = self.step + 1
         if self.step == 1:
-            return self._select_all_mineral_collecter_scv()
+            return self.select_scv(observations)
         elif self.step == 2:
-            return self._build_barracks(observations)
+            return self.build(observations)
+        elif self.step == 3:
+            return self.send_scv_to_mineral(observations)
 
-    def _select_all_mineral_collecter_scv(self) -> actions.FunctionCall:
-        return SCVCommonActions().select_a_group_of_scv(self.scv_groups.mineral_collectors_group_id())
+    def select_scv(self, observations: Observations) -> actions.FunctionCall:
+        return SCVCommonActions().select_scv(observations)
 
-    def _build_barracks(self, observations: Observations) -> actions.FunctionCall:
+    def build(self, observations: Observations) -> actions.FunctionCall:
+        unit_type = observations.screen().unit_type()
         cc_y, cc_x = self.location.command_center_first_position()
-        barracks_count = BuildingCounter().barracks_count(observations)
+        barracks_y, barracks_x = (unit_type == self.unit_type_ids.terran_barracks()).nonzero()
+        barracks_count = int(round(len(barracks_y) / 137))
         if barracks_count < self.max_barracks and self.action_ids.build_barracks() in observations.available_actions():
             if cc_y.any():
-                current_count_to_difference_from_cc = BuildingPositionsFromCommandCenter().barracks()
+                current_count_to_difference_from_cc = [
+                    [15, -10],
+                    [30, -10],
+                    [15, 10],
+                    [30, 10],
+                ]
                 target = self.location.transform_distance(
                     round(cc_x.mean()),
                     current_count_to_difference_from_cc[barracks_count][0],
@@ -249,153 +98,8 @@ class BuildBarrack(SmartOrder):
 
         return self.actions.no_op()
 
-
-class BuildTechLabBarrack(SmartOrder):
-
-    def __init__(self, base_location: Location, barrack_index: int):
-        SmartOrder.__init__(self, base_location)
-        self.step = 0
-        self.scv_groups = SCVControlGroups()
-        self.barrack_index = barrack_index
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 50 and observations.player().vespene() >= 25
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 2
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        if self.step == 0:
-            return self._select_barrack(observations)
-        elif self.step == 1:
-            return self._build_techlab(observations)
-        return self.actions.no_op()
-
-    def _select_barrack(self, observations: Observations) -> actions.FunctionCall:
-        barrack_count = BuildingCounter().barracks_count(observations)
-        if barrack_count >= 1:
-            self.step = self.step + 1
-            cc_y, cc_x = self.location.command_center_first_position()
-            difference_from_cc = self.difference_from_command_center()
-            target = self.location.transform_distance(
-                round(cc_x.mean()),
-                difference_from_cc[0],
-                round(cc_y.mean()),
-                difference_from_cc[1],
-            )
-            return self.actions.select_point(target)
-        return self.actions.no_op()
-
-    def _build_techlab(self, observations: Observations) -> actions.FunctionCall:
-        if self.action_ids.build_techlab_barracks() in observations.available_actions():
-            self.step = self.step + 1
-            cc_y, cc_x = self.location.command_center_first_position()
-            difference_from_cc = self.difference_from_command_center()
-            target = self.location.transform_distance(
-                round(cc_x.mean()),
-                difference_from_cc[0],
-                round(cc_y.mean()),
-                difference_from_cc[1],
-            )
-            return self.actions.build_techlab_barracks(target)
-        return self.actions.no_op()
-
-    def difference_from_command_center(self) -> []:
-        return BuildingPositionsFromCommandCenter().barracks()[self.barrack_index - 1]
-
-
-class BuildReactorBarrack(SmartOrder):
-
-    def __init__(self, base_location: Location, barrack_index: int):
-        SmartOrder.__init__(self, base_location)
-        self.step = 0
-        self.scv_groups = SCVControlGroups()
-        self.barrack_index = barrack_index
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 50 and observations.player().vespene() >= 50
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 2
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        if self.step == 0:
-            return self._select_barrack(observations)
-        elif self.step == 1:
-            return self._build_reactor(observations)
-        return self.actions.no_op()
-
-    def _select_barrack(self, observations: Observations) -> actions.FunctionCall:
-        barrack_count = BuildingCounter().barracks_count(observations)
-        if barrack_count >= 2: # always build on second and third barracks
-            self.step = self.step + 1
-            cc_y, cc_x = self.location.command_center_first_position()
-            difference_from_cc = self.difference_from_command_center()
-            target = self.location.transform_distance(
-                round(cc_x.mean()),
-                difference_from_cc[0],
-                round(cc_y.mean()),
-                difference_from_cc[1],
-            )
-            return self.actions.select_point(target)
-        return self.actions.no_op()
-
-    def _build_reactor(self, observations: Observations) -> actions.FunctionCall:
-        if self.action_ids.build_reactor_barracks() in observations.available_actions():
-            self.step = self.step + 1
-            cc_y, cc_x = self.location.command_center_first_position()
-            difference_from_cc = self.difference_from_command_center()
-            target = self.location.transform_distance(
-                round(cc_x.mean()),
-                difference_from_cc[0],
-                round(cc_y.mean()),
-                difference_from_cc[1],
-            )
-            return self.actions.build_reactor_barracks(target)
-        return self.actions.no_op()
-
-    def difference_from_command_center(self) -> []:
-        return BuildingPositionsFromCommandCenter().barracks()[self.barrack_index - 1]
-
-
-class BuildFactory(SmartOrder):
-
-    def __init__(self, location: Location, max_factories: int = 2):
-        SmartOrder.__init__(self, location)
-        self.step = 0
-        self.max_factories = max_factories
-        self.scv_groups = SCVControlGroups()
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 150 and observations.player().vespene() >= 100
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 2
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        self.step = self.step + 1
-        if self.step == 1:
-            return self._select_all_mineral_collecter_scv()
-        elif self.step == 2:
-            return self._build_factory(observations)
-
-    def _select_all_mineral_collecter_scv(self) -> actions.FunctionCall:
-        return SCVCommonActions().select_a_group_of_scv(self.scv_groups.mineral_collectors_group_id())
-
-    def _build_factory(self, observations: Observations) -> actions.FunctionCall:
-        cc_y, cc_x = self.location.command_center_first_position()
-        factories_count = BuildingCounter().factories_count(observations)
-        if factories_count < self.max_factories and self.action_ids.build_factory() in observations.available_actions():
-            if cc_y.any():
-                current_count_to_difference_from_cc = BuildingPositionsFromCommandCenter().factories()
-                target = self.location.transform_distance(
-                    round(cc_x.mean()),
-                    current_count_to_difference_from_cc[factories_count][0],
-                    round(cc_y.mean()),
-                    current_count_to_difference_from_cc[factories_count][1],
-                )
-                return self.actions.build_factory(target)
-        return self.actions.no_op()
+    def send_scv_to_mineral(self, observations: Observations) -> actions.FunctionCall:
+        return SCVCommonActions().send_scv_to_mineral(observations)
 
 
 class BuildSupplyDepot(SmartOrder):
@@ -404,30 +108,42 @@ class BuildSupplyDepot(SmartOrder):
         SmartOrder.__init__(self, location)
         self.step = 0
         self.max_supplies = max_supplies
-        self.scv_groups = SCVControlGroups()
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 100
 
     def done(self, observations: Observations) -> bool:
-        return self.step == 2
+        return self.step == 3
 
     def execute(self, observations: Observations) -> actions.FunctionCall:
         self.step = self.step + 1
         if self.step == 1:
-            return self._select_all_mineral_collecter_scv()
+            return self.select_scv(observations)
         elif self.step == 2:
-            return self._build_supply_depot(observations)
+            return self.build(observations)
+        elif self.step == 3:
+            return self.send_scv_to_mineral(observations)
 
-    def _select_all_mineral_collecter_scv(self) -> actions.FunctionCall:
-        return SCVCommonActions().select_a_group_of_scv(self.scv_groups.mineral_collectors_group_id())
+    def select_scv(self, observations: Observations) -> actions.FunctionCall:
+        return SCVCommonActions().select_scv(observations)
 
-    def _build_supply_depot(self, observations: Observations) -> actions.FunctionCall:
+    def build(self, observations: Observations) -> actions.FunctionCall:
+        unit_type = observations.screen().unit_type()
         cc_y, cc_x = self.location.command_center_first_position()
-        supply_depot_count = BuildingCounter().supply_depots_count(observations)
+        depot_y, depot_x = (unit_type == self.unit_type_ids.terran_supply_depot()).nonzero()
+        supply_depot_count = int(round(len(depot_y) / 69))
         if supply_depot_count < self.max_supplies and self.action_ids.build_supply_depot() in observations.available_actions():
             if cc_y.any():
-                current_count_to_difference_from_cc = BuildingPositionsFromCommandCenter().supply_depots()
+                # some tweak to make it work on both start positions that seems not symetric
+                current_count_to_difference_from_cc = [
+                    [-35, -20],
+                    [-35, -10],
+                    [-35, 0],
+                    [-35, 10],
+                    [-35, 20],
+                    [-25, -30],
+                    [-15, -30],
+                    [-5, -30],
+                    [5, -35],
+                    [15, -35]
+                ]
                 target = self.location.transform_distance(
                     round(cc_x.mean()),
                     current_count_to_difference_from_cc[supply_depot_count][0],
@@ -437,179 +153,8 @@ class BuildSupplyDepot(SmartOrder):
                 return self.actions.build_supply_depot(target)
         return self.actions.no_op()
 
-
-class BuildRefinery(SmartOrder):
-
-    def __init__(self, base_location: Location, refinery_index: int):
-        SmartOrder.__init__(self, base_location)
-        self.step = 0
-        self.refinery_index = refinery_index
-        self.scv_groups = SCVControlGroups()
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 75
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 2
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        group_id = self._relevant_group_id()
-        self.step = self.step + 1
-        if self.step == 1:
-            return self._select_all_refinery_collecter_scv(group_id)
-        elif self.step == 2:
-            return self._build_refinery(observations)
-        return self.actions.no_op()
-
-    def _relevant_group_id(self) -> int:
-        if self.refinery_index == 1:
-            return self.scv_groups.refinery_one_collectors_group_id()
-        return self.scv_groups.refinery_two_collectors_group_id()
-
-    def _select_all_refinery_collecter_scv(self, group_id: int) -> actions.FunctionCall:
-        return SCVCommonActions().select_a_group_of_scv(group_id)
-
-    def _build_refinery(self, observations: Observations) -> actions.FunctionCall:
-        if self.action_ids.build_refinery() in observations.available_actions():
-            if self.refinery_index == 1:
-                difference_from_cc = BuildingPositionsFromCommandCenter().vespene_geysers()[0]
-            else:
-                difference_from_cc = BuildingPositionsFromCommandCenter().vespene_geysers()[1]
-            cc_y, cc_x = self.location.command_center_first_position()
-            target = self.location.transform_distance(
-                round(cc_x.mean()),
-                difference_from_cc[0],
-                round(cc_y.mean()),
-                difference_from_cc[1],
-            )
-            #print(
-            #    "build refinery "+str(self.refinery_index) + " on "+str(target[0])+ " "+str(target[1])
-            #    + " CC is "+ str(round(cc_x.mean())) + " " + str(round(cc_y.mean()))
-            #)
-
-            return self.actions.build_refinery(target)
-        return self.actions.no_op()
-
-
-class FillRefineryOnceBuilt(SmartOrder):
-
-    def __init__(self, base_location: Location, refinery_index: int):
-        SmartOrder.__init__(self, base_location)
-        self.step = 0
-        self.refinery_index = refinery_index
-        self.scv_groups = SCVControlGroups()
-
-    def doable(self, observations: Observations) -> bool:
-        return True
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 3
-
-    def doable(self, observations: Observations) -> bool:
-        refinery_count = BuildingCounter().refineries_count(observations)
-        return refinery_count >= self.refinery_index
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        if self.step == 0:
-            return self._select_refinery()
-        elif self.step == 1 and self._selected_refinery_is_built(observations):
-            #print("refinery is built, select collectors" + str(self.refinery_index))
-            return self._select_vespene_collectors()
-        elif self.step == 2:
-            return self._send_collectors_to_refinery(observations)
-
-        return self.actions.no_op()
-
-    def _select_refinery(self) -> actions.FunctionCall:
-        self.step = self.step + 1
-        if self.refinery_index == 1:
-            difference_from_cc = BuildingPositionsFromCommandCenter().vespene_geysers()[0]
-        else:
-            difference_from_cc = BuildingPositionsFromCommandCenter().vespene_geysers()[1]
-        cc_y, cc_x = self.location.command_center_first_position()
-        target = self.location.transform_distance(
-            round(cc_x.mean()),
-            difference_from_cc[0],
-            round(cc_y.mean()),
-            difference_from_cc[1],
-        )
-        #print("select refinery " + str(target[0]) + " " + str(target[1]))
-        return self.actions.select_point(target)
-
-    def _selected_refinery_is_built(self, observations: Observations) -> bool:
-        return observations.single_select().is_built()
-
-    def _select_vespene_collectors(self) -> actions.FunctionCall:
-        self.step = self.step + 1
-        if self.refinery_index == 1:
-            group_id = SCVControlGroups().refinery_one_collectors_group_id()
-        else:
-            group_id = SCVControlGroups().refinery_two_collectors_group_id()
-        return SCVCommonActions().select_a_group_of_scv(group_id)
-
-    def _send_collectors_to_refinery(self, observations: Observations) -> actions.FunctionCall:
-        self.step = self.step + 1
-        #print(observations.multi_select())
-        return SCVCommonActions().send_selected_scv_group_to_refinery(self.location, observations, self.refinery_index)
-
-
-class SendIdleSCVToMineral(SmartOrder):
-
-    def __init__(self, base_location: Location):
-        SmartOrder.__init__(self, base_location)
-        self.step = 0
-
-    def doable(self, observations: Observations) -> bool:
-        return True
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 2
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().idle_worker_count() > 0
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        self.step = self.step + 1
-        if self.step == 1:
-            return self.actions.select_idle_worker()
-        elif self.step == 2:
-            return SCVCommonActions().send_scv_to_mineral(observations)
-
-
-class BuildSCV(SmartOrder):
-
-    def __init__(self, location: Location):
-        SmartOrder.__init__(self, location)
-        self.step = 0
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 50
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 2
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().food_used() < observations.player().food_cap()
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        self.step = self.step + 1
-        if self.step == 1:
-            return self._select_command_center(observations)
-        elif self.step == 2:
-            return self._train_scv(observations)
-
-    def _select_command_center(self, observations: Observations) -> actions.FunctionCall:
-        unit_type = observations.screen().unit_type()
-        cc_y, cc_x = (unit_type == self.unit_type_ids.terran_command_center()).nonzero()
-        if cc_y.any():
-            target = [round(cc_x.mean()), round(cc_y.mean())]
-            return self.actions.select_point(target)
-        return self.actions.no_op()
-
-    def _train_scv(self, observations: Observations) -> actions.FunctionCall:
-        if self.action_ids.train_scv() in observations.available_actions():
-            return self.actions.train_scv()
-        return self.actions.no_op()
+    def send_scv_to_mineral(self, observations: Observations) -> actions.FunctionCall:
+        return SCVCommonActions().send_scv_to_mineral(observations)
 
 
 class BuildMarine(SmartOrder):
@@ -617,9 +162,6 @@ class BuildMarine(SmartOrder):
     def __init__(self, location: Location):
         SmartOrder.__init__(self, location)
         self.step = 0
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 50
 
     def done(self, observations: Observations) -> bool:
         return self.step == 2
@@ -646,50 +188,13 @@ class BuildMarine(SmartOrder):
         return self.actions.no_op()
 
 
-class BuildMarauder(SmartOrder):
+class Attack(SmartOrder):
 
-    def __init__(self, location: Location):
-        SmartOrder.__init__(self, location)
-        self.step = 0
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 100 and observations.player().vespene() >= 25
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 2
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        self.step = self.step + 1
-        if self.step == 1:
-            return self.select_barracks(observations)
-        elif self.step == 2:
-            return self.train_marine(observations)
-
-    def select_barracks(self, observations: Observations) -> actions.FunctionCall:
-        unit_type = observations.screen().unit_type()
-        barracks_y, barracks_x = (unit_type == self.unit_type_ids.terran_barracks()).nonzero()
-        if barracks_y.any():
-            i = random.randint(0, len(barracks_y) - 1)
-            target = [barracks_x[i], barracks_y[i]]
-            return self.actions.select_point_all(target)
-        return self.actions.no_op()
-
-    def train_marine(self, observations: Observations) -> actions.FunctionCall:
-        if self.action_ids.train_marauder() in observations.available_actions():
-            return self.actions.train_marauder()
-        return self.actions.no_op()
-
-
-class DumbAttack(SmartOrder):
-
-    def __init__(self, location: Location, x: int , y: int):
+    def __init__(self, location: Location, x , y):
         SmartOrder.__init__(self, location)
         self.step = 0
         self.x = x
         self.y = y
-
-    def doable(self, observations: Observations) -> bool:
-        return True
 
     def done(self, observations: Observations) -> bool:
         return self.step == 2
@@ -718,98 +223,16 @@ class DumbAttack(SmartOrder):
             x_offset = random.randint(-1, 1)
             y_offset = random.randint(-1, 1)
             target = self.location.transform_location(int(self.x) + (x_offset * 8), int(self.y) + (y_offset * 8))
-
             return self.actions.attack_minimap(target)
 
         return self.actions.no_op()
 
 
-# http://liquipedia.net/starcraft2/Tech_Lab_(Legacy_of_the_Void)
-class ResearchCombatShield(SmartOrder):
-
-    def __init__(self, base_location: Location):
-        SmartOrder.__init__(self, base_location)
-        self.step = 0
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 100 and observations.player().vespene() >= 100
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 2
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        if self.step == 0:
-            return self._select_tech_barrack(observations)
-        elif self.step == 1:
-            return self._research_combat_shield(observations)
-        return self.actions.no_op()
-
-    def _select_tech_barrack(self, observations: Observations) -> actions.FunctionCall:
-        techlab_barrack_count = BuildingCounter().techlab_barracks_count(observations)
-        if techlab_barrack_count >= 1:
-            self.step = self.step + 1
-            unit_type = observations.screen().unit_type()
-            center_y, center_x = (unit_type == self.unit_type_ids.terran_barracks_techlab()).nonzero()
-            center_x = round(center_x.mean())
-            center_y = round(center_y.mean())
-            target = [center_x, center_y]
-            return self.actions.select_point(target)
-        return self.actions.no_op()
-
-    def _research_combat_shield(self, observations: Observations) -> actions.FunctionCall:
-        if self.action_ids.research_combat_shield() in observations.available_actions():
-            self.step = self.step + 1
-            return self.actions.research_combat_shield()
-        return self.actions.no_op()
-
-
-# http://liquipedia.net/starcraft2/Tech_Lab_(Legacy_of_the_Void)
-class ResearchConcussiveShells(SmartOrder):
-
-    def __init__(self, base_location: Location):
-        SmartOrder.__init__(self, base_location)
-        self.step = 0
-
-    def doable(self, observations: Observations) -> bool:
-        return observations.player().minerals() >= 50 and observations.player().vespene() >= 50
-
-    def done(self, observations: Observations) -> bool:
-        return self.step == 2
-
-    def execute(self, observations: Observations) -> actions.FunctionCall:
-        if self.step == 0:
-            return self._select_tech_barrack(observations)
-        elif self.step == 1:
-            return self._research_concussive_shells(observations)
-        return self.actions.no_op()
-
-    def _select_tech_barrack(self, observations: Observations) -> actions.FunctionCall:
-        techlab_barrack_count = BuildingCounter().techlab_barracks_count(observations)
-        if techlab_barrack_count >= 1:
-            self.step = self.step + 1
-            unit_type = observations.screen().unit_type()
-            center_y, center_x = (unit_type == self.unit_type_ids.terran_barracks_techlab()).nonzero()
-            center_x = round(center_x.mean())
-            center_y = round(center_y.mean())
-            target = [center_x, center_y]
-            return self.actions.select_point(target)
-        return self.actions.no_op()
-
-    def _research_concussive_shells(self, observations: Observations) -> actions.FunctionCall:
-        if self.action_ids.research_concussive_shells() in observations.available_actions():
-            self.step = self.step + 1
-            return self.actions.research_concussive_shells()
-        return self.actions.no_op()
-
-
-class NoOrder(SmartOrder):
+class NoOrder(Order):
 
     def __init__(self):
-        SmartOrder.__init__(self, None)
+        Order.__init__(self)
         self.step = 0
-
-    def doable(self, observations: Observations) -> bool:
-        return True
 
     def done(self, observations: Observations) -> bool:
         return self.step == 1
